+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This allows to get a more linear FM sound from that oscillator.
 * Change: Changed `Ad` envelope node to be properly retriggerable. Refactored
 out it's DSP code to `synfx_dsp::EnvRetrigAD`.
 * Feature: Added `Matrix::load_patch_from_mem` and `Matrix::save_patch_to_mem`.
+* Feature: The `BOsc` node has now a pulse mode with DC correction.
 
 0.2.0 (2022-08-28)
 ==================